Make trpo work forpendulum

+++ TRPO/policy.py
@@ -33,7 +33,7 @@ class NN(object):
 
     # choose action
     def choose_a(self, s):
-        mu = self.model(torch.tensor(s)).detach().numpy()
+        mu = self.model(torch.tensor(s, dtype=torch.float)).detach().numpy()
         return np.random.multivariate_normal(mu, self.get_covariance_matrix_numpy(), 1)
 
     # pi_theta old
